# openapi.yaml
...
openapi: 3.0.3
info:
  title: rucQ API
  version: 0.1.0
  description: rucQ API
servers:
  - url: http://localhost:8080
    description: 開発環境
paths:
...
  /api/camps/default:
    get:
      summary: デフォルトの合宿を取得
      tags:
        - Camps
      operationId: getDefaultCamp
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camp"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/me/answers/{question_id}:
    get:
      summary: 自分の回答を取得
      tags:
        - Questions
      operationId: getMyAnswer
      parameters:
        - $ref: "#/components/parameters/QuestionId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: 回答を更新
      tags:
        - Questions
      operationId: putAnswer
      parameters:
        - $ref: "#/components/parameters/QuestionId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutAnswerRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/users/{traq_id}/answers/{question_id}:
    get:
      summary: ユーザーの回答を取得
      tags:
        - Questions
      operationId: getUserAnswer
      parameters:
        - $ref: "#/components/parameters/TraqId"
        - $ref: "#/components/parameters/QuestionId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: 回答を更新
      tags:
        - Questions
      operationId: putUserAnswer
      parameters:
        - $ref: "#/components/parameters/TraqId"
        - $ref: "#/components/parameters/QuestionId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutAnswerRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/me/budgets:
    get:
      summary: 自分の予算を取得
      tags:
        - Budgets
      operationId: getMyBudget
      parameters:
        - $ref: "#/components/parameters/X-Forwarded-User"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/users/{traq_id}/budgets:
    get:
      summary: ユーザーの予算を取得
      tags:
        - Budgets
      operationId: getUserBudget
      parameters:
        - $ref: "#/components/parameters/TraqId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: ユーザーの予算を設定
      tags:
        - Budgets
      operationId: postUserBudget
      parameters:
        - $ref: "#/components/parameters/TraqId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostBudgetRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: ユーザーの予算を更新
      tags:
        - Budgets
      operationId: putUserBudget
      parameters:
        - $ref: "#/components/parameters/TraqId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostBudgetRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/budgets:
    get:
      summary: 支払い情報の一覧を取得
      tags:
        - Budgets
      operationId: getBudgets
      parameters:
        - $ref: "#/components/parameters/CampIdQuery"
        - $ref: "#/components/parameters/X-Forwarded-User"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    X-Forwarded-User:
      name: X-Forwarded-User
      in: header
      description: ログインしているユーザーのtraQ ID（NeoShowcaseが自動で付与）
      schema:
        type: string
    TraqId:
      name: traq_id
      in: path
      description: traQ ID
      required: true
      schema:
        type: string
...
    CampIdQuery:
      name: camp_id
      in: query
      description: 合宿ID
      schema:
        type: integer
...
    QuestionId:
      name: question_id
      in: path
      description: 質問ID
      required: true
      schema:
        type: integer
...
  responses:
...
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Internal server error
  schemas:
    Camp:
      type: object
      properties:
        id:
          type: integer
        display_id:
          type: string
        name:
          type: string
        is_draft:
          type: boolean
        description:
          type: string
      required:
        - id
        - display_id
        - name
        - is_draft
        - description
...
    Answer:
      type: object
      properties:
        id:
          type: integer
        question_id:
          type: integer
        user_traq_id:
          type: string
        content:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          nullable: true
      required:
        - id
        - question_id
        - user_traq_id
    PutAnswerRequest:
      type: object
      properties:
        content:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          nullable: true
    Budget:
      type: object
      properties:
        id:
          type: integer
        user_traq_id:
          type: string
        camp_id:
          type: integer
        amount:
          type: integer
          nullable: true
        amount_paid:
          type: integer
      required:
        - id
        - user_traq_id
        - camp_id
        - amount_paid
    PostBudgetRequest:
      type: object
      properties:
        camp_id:
          type: integer
        amount:
          type: integer
          nullable: true
        amount_paid:
          type: integer
      required:
        - camp_id
        - amount_paid
...